requirements.txt
# Core ML + Transformers
torch==2.8.0
transformers==4.44.0
numpy
scipy

# Web frameworks
flask
streamlit

# Spotify API
spotipy

# Utilities
requests
pandas
python-dotenv   # ✅ new addition for environment variable support
